Add the concept of batching requests
This introduces the concept of batches in the database
and REST API. This paves the way for a future change
that will allow a user to submit multiple requests
at the same time and keep track of them as a batch.

All existing requests will be individually assigned to
a new batch as part of the database migration. All new
requests will be individually assigned to a new batch.