Overwrite from_index before the OMPS push [CLOUDDST-114]
Earlier if overwrite_from_index option was specified,
the index image would be overridden after the OMPS push.
Fix this as overriding should happen before and
the OMPS push should get the latest pull spec of the index image.
Also so that the OMPS push can be ignored if index override
raised any errors