Use skopeo copy for pushing image
Podman tries to determine which manifest format should be used when
pushing an image. If the registry supports the OCI format, it'll always
use that. However, some registries do not support this format yet. If
the image was pushed to the IIB registry in the OCI format, skopeo
cannot copy that image to another registry that does not support the OCI
format. This poses a limitation when IIB must overwrite the from index
image, and usage of IIB produced images in general.

Because skopeo allows controlling the exact manifest format, it can be
used to force the more compatible format v2s2.

This change also bumps the default skopeo timeout to 300s to ensure the
image can be fully copied.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>

fixup! Use skopeo copy for pushing image