Add support for overwriting index images that use the same registry as IIB
When an index image is built, IIB always pushes it to its configured
registry. Additionally, if `overwrite_from_index` is provided, IIB
will copy the index image from its registry over to the location of
`from_index`.

If `overwrite_from_index_token` is provided and `from_index` is
on the same registry as IIB's configured registry, then this will
fail since the `overwrite_from_index_token` token does not have
access to the source repository to copy from.

In this case, this change makes IIB first export the index image
in IIB's configured registry and repo to local OCI images using
its token and then push the local OCI images using the
`overwrite_from_index_token` token.

When buildah is updated in RHEL 8 and manifest-tool is no longer
needed, this workaround can be removed.

Resolves CLOUDBLD-1502