Add merge-index-image endpoint
The endpoint takes in a source_index_image which is used as a base
to create a new index image with the ocp_version of the target_index
parameter. It also adds any additional bundles present in target_index
but not in source_from_index. Apart from that, a deprecation_list can be
provided to deprecate the bundles in the resulting index image.

Refers to CLOUDDST-2174 and CLOUDDST-2175

Co-authored-by: Yashvardhan Nanavati <yashn@bu.edu>