Use podman container-tool when merging indexes
opm has a built-in mechanism for interacting with container image. This
is the default. However, when adding large amount of bundle image to an
index image, opm's built-in mechanism does not handle availability
hiccups in the registry hosting the bundle images:
https://bugzilla.redhat.com/show_bug.cgi?id=1937097

In fact, we have not been able to run a merge index operation
successfully when adding a couple of hundred bundle images to the index.
Some of the indexes managed by IIB have double this amount.

By setting container-tool to podman when adding images during a merge
operation, we drastically improve the resiliency of this operation.

* CLOUDBLD-4658

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>