Set correlation_id, not message_id, on messages
The ActiveMQ broker will set a message ID automatically if one is not
set. This is actually preferred as it allows custom configuration in the
broker to be leveraged to created a meaningful message ID.

For tracking purposes, it is encouraged to set a correlation_id to allow
tracking of the message throughout the system.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>