Add support for AWS S3 buckets for artifacts storage
User can now choose if they want to utilize local storage or AWS S3 buckets
for storing artifacts like logs. For iib-web, local storage is optional when
S3 bucket is configured. For iib-workers, local storage is mandatory when
S3 bucket is configured. This is because IIB uploads the files to the S3 bucket
from local storage. Existing functionality of only using local storage for
artifacts remains intact.

Refers to CLOUDDST-7333