Remove operator packages from a FBC index image.
This commit is for removing the specified operator from
FBC index image. The image is identified with the label
"operators.operatorframework.io.index.configs.v1".
If the index image is FBC, the operator is removed
by using opm registry commands and then index is migrated
to FBC format, generating the dockerfile and building the
resultant image in the end.

Refers to CLOUDDST-8616