Fix loosing olm.maxOpenShiftVersion property
The issue is that after adding olm.maxOpenShiftVersion property
directly to index.db we did not build new image.
When merge API was called and deprecation_bundles were set
the function deprecate_bundles worked with intermediate image
that was pushed to external registry. Unfortunately it does not
contain olm.maxOpenShiftVersion since no image with this change
was never built and pushed to that registry.

Fix - build and push image with olm.maxOpenShiftVersion property