Add internal_index_image_copy and internal_index_image_copy_resolved attributes
Added attributes to response of Add and Rm requests. Their values will be the
pullspec pointing to the default IIB output repo. The values will be same as
that of index_image and index_image_resolved respectively when overwrite_from_index
is NOT provided when submitting the request.

Refers to CLOUDDST-13288