Log a warning message when "architecture" label is missing under "Labels"
When using an Openshift binary as the binary_image, the "architecture" label
is added under the "Labels" section on an index image. This is done for the
OCP cluster to recognize the arch of the image. This is an OCP specific thing
and it's not necessary that the "architecture" label will be added under
"Labels" if a user specifies any other binary_image. Hence, in-case the label
is missing from the index, do not fail the request.

Refers to CLOUDDST-19860