Fix normalization bug when return value is non-dict
The changes added in #557 assumed the return value of a function
will be a dictionary. In the current instrumentation, _get_catalog_dir
is the exception and returns the string in which case we need the change
that was originally present. This PR combines the two changes and allows
the decorator to process a non-dict return_value.