Add guardrails for UniqueViolation caused by parallel requests
Sometime, when multiple requests are submitted with the same
parameters, it can cause a UniqueViolation in the database.
Example, when requests A and B started, a value wasn't present so they
both decide to add a value X. A's addition is successful while when
B tries to add it, a UniqueViolation is triggered since A already
added it.

Refers to CLOUDDST-20863