Fix memory leak in `opm_render`
This commit aims to fix a memory leak on `opm_render` caused by
returning a list of JSON elements whenever the render operations results
in a voluptous amount of operators.

The adopted solution was to simply tranform the returned list into a
generator which will return chunks of the render data in order to
prevent the memory leak.

Relates to #784

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>