bugfix(workers): Only save to git FBC images
This commit fixes a bug on `_overwrite_from_index` which was always
trying to store the `/configs` directory whenever an overwrite from
index operation happens for `add` or `rm` requests.

The issue is that whenever an image is legacy (SQLite DB) it would fail,
preventing the index images to being overwritten.

The fix simply adds a check condition to inform whether the desired
image to be overwritten is FBC or not.

Refers to CLOUDDST-28815

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>