bugfix(git_utils): Use the git_url_map from params
This commit changes the `git_utils.py` to receive the index image to git
URL mapping as a parameter, since this value isn't stored on IIB workers
config.

It also adjusts the methods `handle_[add|rm]_request` from `build.py` to
pass the index to git url mappings as parameter for the lower functions.

Refers to CLOUDDST-28805

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>