fix(git_utils): get_git_token format adjustment
This commit fixes a bug on `get_git_token` as the
`iib_index_configs_gitlab_tokens_map` will always be a `dict[str, str]`
from HashiVault, never being able to be a Tuple.

In order to return a Tuple we need to split the token-name from
token-value. For that, we're stablishing the standard as a column `:` to
separate such values.

Refers to CLOUDDST-29117

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>