merge_index_image: sort bundles for deprecation
This commit introduces a new function for
`iib.workers.tasks.build_merge_index_image` named
`sort_bundles_by_semver` which receives a list of bundles pullspecs to
sort and a list of bundle images to retrieve their respective versions.
This new function sorts the bundle pullspecs by their operator version
and return the sorted pullspec list.

It's used on `handle_merge_request` to sort the deprecation bundles list
by their `semver` version in order to avoid errors during the
`opm deprecatetruncate` execution that might happen if a major version
deprecates a minor on before it is presented in the list.

Refers to CLOUDDST-28807

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>